refactor: ManagePage 으로 alert 로직 추출
