+++ src/js/constants/index.js
@@ -1,0 +1,4 @@
+const MIN_INPUT = 10;
+const MAX_INPUT = 50000;
+
+export { MIN_INPUT, MAX_INPUT };
