+++ src/js/model/VendingMachine.js
@@ -56,44 +56,12 @@ class VendingMachine {
   }
 
   addProduct({ name, price }) {
-    // TODO 공백도 허용하지 않는다고 적어놔야할듯
-    // TODO: 15라인 이내로 줄이기
-    // TODO: 정규식 상수화 하기
-    // TODO: 모델에 존재해기 or 페이지 컴포넌트로 빼기
-    if (!new RegExp(/^[가-힣]{2,20}$/).test(name)) {
-      window.alert('잘못된 형식의 상품이름입니다.');
-
-      return;
-    }
-    if (!isValidMoneyInput(Number(price))) {
-      window.alert('가격은 10원이상 50000원 이하여야 합니다.');
-
-      return;
-    }
-    if (this.products.map(({ name }) => name).includes(name)) {
-      window.alert('중복된 상품입니다.');
-
-      return;
-    }
-
     this.products.push({ name, price });
 
     setLocalStorageItem('products', this.products);
   }
 
-  removeProduct(name) {
-    const targetIndex = this.products.findIndex(
-      (product) => product.name === name
-    );
-
-    if (targetIndex === -1) {
-      window.alert('존재하지 않는 상품입니다.');
-
-      return;
-    }
-
-    if (!window.confirm('정말로 삭제하시겠습니까?')) return;
-
+  removeProduct(targetIndex) {
     this.products.splice(targetIndex, 1);
     setLocalStorageItem('products', this.products);
   }
