+++ src/js/utils/index.js
@@ -1,5 +1,10 @@
 /* eslint-disable consistent-return */
+
+import { MAX_INPUT, MIN_INPUT } from '../constants/index.js';
+
 const $ = (selector, target = document) => target.querySelector(selector);
+
+const isValidMoneyInput = (value) => MIN_INPUT <= value && value <= MAX_INPUT;
 
 // TODO: 유효성 검사 필요
 const setLocalStorageItem = (key, value) => {
@@ -17,4 +22,4 @@ const getLocalStorageItem = (key) => {
   }
 };
 
-export { $, setLocalStorageItem, getLocalStorageItem };
+export { $, setLocalStorageItem, getLocalStorageItem, isValidMoneyInput };
